==> 7_wikimetrics.ipynb <==
## Summary

This notebook generates a first dataset of universities with bibliometric and Wikipedia data.

# Packages

import pandas as pd

# Wikimetrics

df_m = pd.read_csv('data/wikimetrics.tsv', sep='\t', encoding='UTF-8')
df_m

df_lg = pd.read_csv('data/langlinks.tsv', sep='\t', encoding='UTF-8')
df_lg

df_lg[df_lg.url.duplicated()]

df_m[df_m['url'].isin(df_lg.url.tolist())]

df_m = df_lg[['id', 'url']].drop_duplicates().merge(df_m, how='inner', on='url')
df_m

df_m = df_m[['id', 'editors', 'revisions', 'words', 'unique_references', 'minor_edits', 'characters', 'sections', 'references']]
df_m

df_m = df_m.groupby('id').sum().reset_index()
df_m

df_l = pd.read_csv('data/leiden_wikipedia_indicators.tsv', sep='\t', encoding='UTF-8')
df_l

df_l = df_l.merge(df_m, how='inner', on='id')
df_l

df_l.to_csv('data/full_leiden_wikipedia_indicators.tsv', sep='\t', index=False)
