==> 9_analysis.ipynb <==
## Summary

This notebook generates a table of aggregate metrics per country.

import pandas as pd

# Packages

df = pd.read_csv('data/full_leiden_wikipedia_spec_indicators.tsv', sep='\t', encoding='UTF-8')
df

# Summary table

df_st = df[['Country', 'impact_P', 'impact_P_Bio_Health', 'impact_P_Life_Earth', 'impact_P_Math_Comp',
            'impact_P_Phy_Eng', 'impact_P_Soc_Hum', 'P_top10', 'collab_P','P_int_collab', 'P_OA', 'TCS',
            'TNCS', 'P_industry_collab', 'PA_F_MF', 'dimensions_wos', 'langlinks', 'total_views', 'local_views',
            'int_views', 'rest_views', 'editors', 'revisions', 'words', 'unique_references', 'minor_edits',
            'characters', 'sections', 'references', 'topic_specialisation', 'topic_specialisation_bio',
            'topic_specialisation_eng', 'gini_production', 'established_year', 'age']].copy()
df_st

df_st = df_st.groupby('Country').mean().round(2).reset_index()
df_st

df_st = df_st.merge(df[['Country']].groupby('Country').size().reset_index(name='Universities'), how='inner', on='Country')
df_st.sort_values(by='Universities', ascending=False)

df_st.to_csv('results/agg_values.tsv', sep='\t', index=False)
